refactor: update CourseObject and derived classes to use numeric types for coordinates enhance: add duplicate detection thresholds and optimize point cloud processing in Sensors

+++ mhsboat_ctrl/course_objects.py
@@ -1,10 +1,11 @@
 from typing import Literal
 
+from mhsboat_ctrl.utils.custom_types import numeric
 from mhsboat_ctrl.enums import BuoyColors, Shapes
 
 
 class CourseObject:
-    def __init__(self, x: int, y: int, z: int):
+    def __init__(self, x: numeric, y: numeric, z: numeric):
         self.x = x
         self.y = y
         self.z = z
@@ -15,7 +16,7 @@ class CourseObject:
 
 
 class Shape(CourseObject):
-    def __init__(self, x: int, y: int, z: int, shape: Shapes, color: BuoyColors):
+    def __init__(self, x: numeric, y: numeric, z: numeric, shape: Shapes, color: BuoyColors):
         self.x = x
         self.y = y
         self.z = z
@@ -28,7 +29,7 @@ class Shape(CourseObject):
 
 
 class Buoy(CourseObject):
-    def __init__(self, x: int, y: int, z: int, color: BuoyColors):
+    def __init__(self, x: numeric, y: numeric, z: numeric, color: BuoyColors):
         self.x = x
         self.y = y
         self.z = z
@@ -41,7 +42,7 @@ class Buoy(CourseObject):
 
 class PoleBuoy(Buoy):
     def __init__(
-        self, x: int, y: int, z: int, color: Literal[BuoyColors.RED, BuoyColors.GREEN]
+        self, x: numeric, y: numeric, z: numeric, color: Literal[BuoyColors.RED, BuoyColors.GREEN]
     ):
         self.x = x
         self.y = y
@@ -54,7 +55,7 @@ class PoleBuoy(Buoy):
 
 
 class BallBuoy(Buoy):
-    def __init__(self, x: int, y: int, z: int, color: BuoyColors):
+    def __init__(self, x: numeric, y: numeric, z: numeric, color: BuoyColors):
         self.x = x
         self.y = y
         self.z = z
